Create Tip calculator project
we created a calculator by adding a tip in which when a group of friends went to have a lunch or dinner to split the bill this calculator can be used. Here we provided  $10, $12 and $15 as tip we can give tip however we want. This gives the output bill in which how much a single person should pay to the corresponding bill.